Revert "Merge branch 'master' into dev"
This reverts commit f971fb9016d45404370f0a68787e68b8da6ed212, reversing
changes made to dc4e26ee51a898885d7b400add360d398b3f5cbf.
