+++ .eslintrc.js
@@ -22,10 +22,9 @@ module.exports = {
     '@typescript-eslint/type-annotation-spacing': 'off',
     '@typescript-eslint/array-type': 'off',
     '@typescript-eslint/return-await': 'off',
-    '@typescript-eslint/parser': 'off',
     'import/no-extraneous-dependencies': ['error', { devDependencies: ['**/*.spec.ts'] }],
     'react-hooks/rules-of-hooks': 'error',
-    'import/no-unresolved': 'off',
+    "import/no-unresolved": "off",
     'no-return-await': 'off',
     'react-hooks/exhaustive-deps': [
       'off',
