+++ src/language/en-US.json
@@ -131,7 +131,7 @@
   "general.table.triggerDesc": "Click to sort by descend",
   "general.table.triggerAsc": "Click to sort by ascend",
   "general.table.cancelSort": "Click to cancel sort",
-  "welcome.slogan": "A public, open-source Desktop Wallet on Crypto.org Chain.",
+  "welcome.slogan": "A public, open-source Destkop Wallet on Crypto.org Chain.",
   "welcome.button": "Get Started",
   "welcome.block": "This product is currently unavailable in your country or region.",
   "signup.title": "App Password Created Successfully",
@@ -342,7 +342,6 @@
   "bridge.form.fee": "Fee: ",
   "bridge.form.destination": "Destination: ",
   "bridge.form.toAddress": "To Address: ",
-  "bridge.form.customAddress.message": "I want to send funds to a custom destination address.",
   "bridge.form.transfer": "Transfer Asset",
   "bridge.form.receiving": "Receiving",
   "bridge.form.disclaimer": "By proceeding, I hereby acknowledge that I agree to the terms and conditions.",
@@ -374,7 +373,6 @@
   "bridge.transferInitiated.description": "Check your wallet for the funds. It may take up to a few minutes depending on network congestion.",
   "bridge.pendingTransfer.title": "{{amount}} {{symbol}} is pending to transfer",
   "bridge.pendingTransfer.ledger.description": "Please open {{app}} in your Ledger device and sign your transaction.",
-  "bridge.pendingTransferTimeout.description": "It may take up to a few minutes depending on network congestion.",
   "bridge.ledgerSign.failed.title": "Failed to broadcast a signed transaction on Ledger",
   "bridge.transactionHistory.title": "Bridge Transfer History",
   "bridge.transactionHistory.description": "Below is your latest transaction history of bridge transfers",
