+++ src/language/ko-KR.json
@@ -76,8 +76,6 @@
   "general.sessionLockModal.errorSelfDestruct": "#*N*을 시도합니다.  #T#번  잘못된 시도를 하면 지갑은 자동 소멸됩니다.",
   "general.sessionLockModal.success": "지갑을 성공적으로 해독하였습니다",
   "general.sessionLockModal.okButton": "지갑 해독",
-  "general.sessionLockModal.notification.message1": "앱 잠김",
-  "general.sessionLockModal.notification.description1": "앱이 곧 잠깁니다.",
   "general.sessionLockModal.notification.message2": "지갑 해독",
   "general.sessionLockModal.notification.description2": "지갑 해독 성공",
   "general.successModalPopup.title": "성공!",
@@ -340,7 +338,6 @@
   "bridge.form.fee": "수수료",
   "bridge.form.destination": "목적지",
   "bridge.form.toAddress": "수신인 주소",
-  "bridge.form.customAddress.message": "지정된 주소로 송금하고 싶습니다",
   "bridge.form.transfer": "전송 자산",
   "bridge.form.receiving": "수신 중",
   "bridge.form.disclaimer": "계속 진행함으로써 이용 약관에 동의함을 인정합니다.",
@@ -372,7 +369,6 @@
   "bridge.transferInitiated.description": "지갑에서 자금을 확인하십시오. 네트워크 상태에 따라 최대 몇분까지 소요될 수 있습니다.",
   "bridge.pendingTransfer.title": "{{amount}} {{symbol}}이 전송 대기 중입니다",
   "bridge.pendingTransfer.ledger.description": "레져에서 {{app}}을 열고 거래에 서명하세요.",
-  "bridge.pendingTransferTimeout.description": "네트워크 상태에 따라 최대 몇분까지 소요될 수 있습니다.",
   "bridge.ledgerSign.failed.title": "레져에서 서명된 트랜잭션을 전송하지 못했습니다.",
   "bridge.transactionHistory.title": "브릿지 거래 기록",
   "bridge.transactionHistory.description": "다음은 최신 브릿지 거래 내역입니다",
