+++ src/language/zh-CN.json
@@ -342,7 +342,6 @@
   "bridge.form.fee": "费用：",
   "bridge.form.destination": "目标地址：",
   "bridge.form.toAddress": "收款人地址：",
-  "bridge.form.customAddress.message": "我想转移至另一自定收款人地址",
   "bridge.form.transfer": "转移资产",
   "bridge.form.receiving": "将会收到",
   "bridge.form.disclaimer": "我同意条款和条件。",
@@ -374,7 +373,6 @@
   "bridge.transferInitiated.description": "请查看钱包资产。在网络拥堵的情况下，可能最多需要数分钟完成交易。",
   "bridge.pendingTransfer.title": "{{amount}} {{symbol}} 正等待发送",
   "bridge.pendingTransfer.ledger.description": "请在Ledger钱包上打开 {{app}} 以签名交易",
-  "bridge.pendingTransferTimeout.description": "在网络拥堵的情况下，可能最多需要数分钟完成交易。",
   "bridge.ledgerSign.failed.title": "未能成功广播Ledger的签名交易",
   "bridge.transactionHistory.title": "跨链桥交易记录",
   "bridge.transactionHistory.description": "以下是你最近的跨链桥交易记录",
