+++ src/models/Transaction.ts
@@ -151,8 +151,6 @@ export interface ValidatorModel {
   readonly currentShares: string;
   readonly pubKey: ValidatorPubKey;
   displayWarningColumn?: boolean;
-  apy?: string;
-  uptime?: string;
   cumulativeSharesIncludePercentage?: string;
   cumulativeSharesExcludePercentage?: string;
 }
