+++ src/pages/assets/components/FormSend.less
@@ -34,17 +34,4 @@
     margin: 0 !important;
   }
   //   }
-
-  .address-name-label {
-    position: relative;
-    top: -15px;
-    color: @font-color;
-    text-align: left;
-  }
-
-  .add-to-address-list-checkbox {
-    .ant-form-item-control-input {
-      min-height: 0;
-    }
-  }
 }
