+++ src/pages/bridge/components/CronosBridgeForm.less
@@ -37,8 +37,5 @@
     .ant-form-item-explain-error {
       text-align: left;
     }
-    .ant-form-item-has-feedback .ant-input {
-      padding-right: 44px;
-    }
   }
 }
