+++ src/pages/bridge/components/CronosBridgeForm.tsx
@@ -1,11 +1,10 @@
 import React, { useEffect, useRef, useState } from 'react';
 import './CronosBridgeForm.less';
-import { Button, Checkbox, Form, Input, InputNumber, Select } from 'antd';
+import { Button, Form, InputNumber, Select } from 'antd';
 import { SwapOutlined } from '@ant-design/icons';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import Big from 'big.js';
 import { useTranslation } from 'react-i18next';
-import { AddressType } from '@crypto-org-chain/chain-jslib/lib/dist/utils/address';
 import { sessionState, walletAllAssetsState } from '../../../recoil/atom';
 import { scaledBalance, UserAsset, UserAssetType } from '../../../models/UserAsset';
 import { middleEllipsis, getCryptoOrgAsset, getCronosAsset } from '../../../utils/utils';
@@ -81,7 +80,6 @@ const CronosBridgeForm = props => {
   const [bridgeSupportedAssets, setBridgeSupportedAssets] = useState<UserAsset[]>([]);
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const [isButtonLoading, setIsButtonLoading] = useState(false);
-  const [isToAddressDisabled, setIsToAddressDisabled] = useState(true);
   const didMountRef = useRef(false);
 
   const analyticsService = new AnalyticsService(session);
@@ -93,12 +91,6 @@ const CronosBridgeForm = props => {
   const { tendermintAddress, evmAddress } = formValues;
 
   const [t] = useTranslation();
-
-  const customAddressValidator = TransactionUtils.addressValidator(
-    session,
-    toAsset,
-    AddressType.USER,
-  );
 
   const getBridgeSupportedAssetList = (assetType: UserAssetType) => {
     return walletAllAssets.filter(asset => {
@@ -115,7 +107,10 @@ const CronosBridgeForm = props => {
     setCurrentAsset(undefined);
     setCurrentAssetIdentifier(undefined);
     setAvailableBalance('--');
-    setIsToAddressDisabled(true);
+    form.setFieldsValue({
+      asset: undefined,
+      amount: undefined,
+    });
 
     switch (bridgeFrom) {
       case 'CRYPTO_ORG': {
@@ -143,34 +138,19 @@ const CronosBridgeForm = props => {
 
     switch (bridgeTo) {
       case 'CRYPTO_ORG': {
-        form.setFieldsValue({
-          toAddress: tendermintAddress,
-        });
         setToAddress(tendermintAddress);
         setToAsset(croAsset);
         break;
       }
       case 'CRONOS': {
-        form.setFieldsValue({
-          toAddress: evmAddress,
-        });
         setToAddress(evmAddress);
         setToAsset(cronosAsset);
         break;
       }
       default: {
-        form.setFieldsValue({
-          toAddress: tendermintAddress,
-        });
         setToAddress(tendermintAddress);
       }
     }
-
-    form.setFieldsValue({
-      asset: undefined,
-      amount: undefined,
-      isCustomToAddress: false,
-    });
   };
 
   const onBridgeExchange = () => {
@@ -205,35 +185,23 @@ const CronosBridgeForm = props => {
 
     switch (newBridgeTo) {
       case 'CRYPTO_ORG': {
-        form.setFieldsValue({
-          toAddress: tendermintAddress,
-        });
         setToAddress(tendermintAddress);
         setToAsset(croAsset);
         break;
       }
       case 'CRONOS': {
-        form.setFieldsValue({
-          toAddress: evmAddress,
-        });
         setToAddress(evmAddress);
         setToAsset(cronosAsset);
         break;
       }
       default: {
-        form.setFieldsValue({
-          toAddress: tendermintAddress,
-        });
         setToAddress(tendermintAddress);
       }
     }
 
-    setIsToAddressDisabled(true);
-
     form.setFieldsValue({
       bridgeFrom: newBridgeFrom,
       bridgeTo: newBridgeTo,
-      isCustomToAddress: false,
     });
     form.validateFields();
   };
@@ -275,11 +243,8 @@ const CronosBridgeForm = props => {
       });
       setSupportedBridges(bridges);
 
-      const { bridgeFrom, bridgeTo, amount, isCustomToAddress } = form.getFieldsValue();
+      const { bridgeFrom, bridgeTo, amount } = form.getFieldsValue();
       if (bridgeFrom && bridgeTo && amount) {
-        if (isCustomToAddress) {
-          setIsToAddressDisabled(false);
-        }
         setAvailableBalance(scaledBalance(currentAsset!));
         setIsBridgeValid(true);
         setSendingAmount(amount);
@@ -318,7 +283,6 @@ const CronosBridgeForm = props => {
 
         form.setFieldsValue({
           asset: cro?.identifier,
-          toAddress: cronos?.address,
         });
         setFormValues({
           tendermintAddress: cro?.address,
@@ -517,12 +481,7 @@ const CronosBridgeForm = props => {
         <div className="ant-row ant-form-item" style={{ marginBottom: '8px' }}>
           {' '}
         </div>
-        <RowAmountOption
-          form={form}
-          walletAsset={currentAsset}
-          setSendingAmount={setSendingAmount}
-          style={{ marginBottom: '8px' }}
-        />
+        <RowAmountOption form={form} walletAsset={currentAsset} style={{ marginBottom: '8px' }} />
       </div>
       <div className="row">
         <div className="ant-row ant-form-item"> </div>
@@ -533,73 +492,6 @@ const CronosBridgeForm = props => {
           </div>
         </div>
       </div>
-      <div className="row">
-        <div className="ant-row ant-form-item">
-          <Form.Item name="isCustomToAddress">
-            <Checkbox
-              checked={!isToAddressDisabled}
-              onChange={e => {
-                const values = form.getFieldsValue();
-                form.setFieldsValue({
-                  ...values,
-                  isCustomToAddress: e.target.checked,
-                });
-                setIsToAddressDisabled(!isToAddressDisabled);
-                if (!e.target.checked) {
-                  const { bridgeTo } = values;
-
-                  switch (bridgeTo) {
-                    case 'CRYPTO_ORG': {
-                      setToAddress(croAsset?.address);
-                      form.setFieldsValue({
-                        toAddress: croAsset?.address,
-                      });
-                      break;
-                    }
-                    case 'CRONOS': {
-                      setToAddress(cronosAsset?.address);
-                      form.setFieldsValue({
-                        toAddress: cronosAsset?.address,
-                      });
-                      break;
-                    }
-                    default:
-                  }
-                }
-              }}
-              className="disclaimer"
-            >
-              {t('bridge.form.customAddress.message')}
-            </Checkbox>
-          </Form.Item>
-        </div>
-        <div className="ant-row ant-form-item">
-          <Form.Item
-            name="toAddress"
-            // label="Destination Address"
-            hasFeedback
-            validateFirst
-            rules={[
-              {
-                required: true,
-                message: `${t('send.formSend.recipientAddress.label')} ${t('general.required')}`,
-              },
-              customAddressValidator,
-            ]}
-            initialValue={toAddress}
-          >
-            <Input
-              placeholder={t('send.formSend.recipientAddress.placeholder')}
-              disabled={isToAddressDisabled}
-              value={toAddress}
-              onChange={event => {
-                setToAddress(event.target.value);
-              }}
-            />
-          </Form.Item>
-        </div>
-      </div>
-
       {currentAsset && new Big(sendingAmount).gt(0) ? (
         <div className="review-container">
           <div className="flex-row">
