+++ src/service/bridge/BridgeService.ts
@@ -68,9 +68,7 @@ export class BridgeService {
   }
 
   private async handleCronosToCryptoOrgTransfer(bridgeTransferRequest: BridgeTransferRequest) {
-    const { originAsset, isCustomToAddress, toAddress, tendermintAddress } = bridgeTransferRequest;
-
-    const recipientAddress = isCustomToAddress ? toAddress : tendermintAddress;
+    const { originAsset } = bridgeTransferRequest;
 
     if (!originAsset.config?.nodeUrl || !originAsset.address) {
       throw TypeError(`Missing asset config: ${originAsset.config}`);
@@ -85,7 +83,7 @@ export class BridgeService {
 
     const txConfig: TransactionConfig = {
       from: bridgeTransferRequest.evmAddress,
-      to: recipientAddress,
+      to: bridgeTransferRequest.tendermintAddress,
       value: web3.utils.toWei(bridgeTransferRequest.amount, 'ether'),
     };
 
@@ -107,7 +105,9 @@ export class BridgeService {
     const gasLimit = loadedBridgeConfig.gasLimit || defaultBridgeConfig.gasLimit;
 
     const contract = new web3.eth.Contract(bridgeContractABI, bridgeContractAddress);
-    const encodedABI = contract.methods.send_cro_to_crypto_org(recipientAddress).encodeABI();
+    const encodedABI = contract.methods
+      .send_cro_to_crypto_org(bridgeTransferRequest.tendermintAddress)
+      .encodeABI();
 
     const scaledBaseAmount = getBaseScaledAmount(bridgeTransferRequest.amount, originAsset);
 
@@ -199,9 +199,7 @@ export class BridgeService {
     );
 
     const evmToBech32ConvertedRecipient = getBech32AddressFromEVMAddress(
-      bridgeTransferRequest.isCustomToAddress
-        ? bridgeTransferRequest.toAddress
-        : bridgeTransferRequest.evmAddress,
+      bridgeTransferRequest.evmAddress,
       loadedBridgeConfig?.prefix || defaultBridgeConfig.prefix,
     );
