+++ src/service/rpc/ChainIndexingAPI.ts
@@ -14,7 +14,6 @@ import {
   accountMsgList,
   ValidatorListResponse,
   AccountInfoResponse,
-  ValidatorResponse,
 } from './ChainIndexingModels';
 import {
   NftQueryParams,
@@ -296,25 +295,7 @@ export class ChainIndexingAPI implements IChainIndexingAPI {
     return listedValidatorInfo;
   }
 
-  public async getValidatorsDetail(validatorAddrList: string[]) {
-    const validatorList = await this.axiosClient.get<ValidatorListResponse>(
-      `validators?limit=1000000`,
-    );
-
-    if (validatorList.data.pagination.total_page > 1) {
-      throw new Error('Validator list is very big. Aborting.');
-    }
-
-    // Check if returned list is empty
-    if (validatorList.data.result.length < 1) {
-      return [];
-    }
-    return validatorList.data.result.filter(validatorInfo =>
-      validatorAddrList.includes(validatorInfo.operatorAddress),
-    );
-  }
-
-  public async getValidatorsAverageApy(validatorAddrList: string[]) {
+  private async getValidatorsAverageApy(validatorAddrList: string[]) {
     const validatorList = await this.axiosClient.get<ValidatorListResponse>(
       `validators?limit=1000000`,
     );
@@ -338,24 +319,7 @@ export class ChainIndexingAPI implements IChainIndexingAPI {
     });
 
     // Listed ValidatorInfo
-    return apySum.div(listedValidatorInfo.length || 1).toString();
-  }
-
-  // NOTE: getting validator by address doesn't have `apy` property
-  public async getValidatorUptimeByAddress(validatorAddr: string) {
-    const validatorInfo = await this.axiosClient.get<ValidatorResponse>(
-      `validators/${validatorAddr}`,
-    );
-
-    if (!validatorInfo.data.result) {
-      throw new Error('Validator details not found.');
-    }
-
-    if (validatorInfo.data.result && !validatorInfo.data.result.impreciseUpTime) {
-      return '0';
-    }
-
-    return validatorInfo.data.result.impreciseUpTime;
+    return apySum.div(listedValidatorInfo.length).toString();
   }
 
   /**
