+++ src/service/rpc/ChainIndexingModels.ts
@@ -190,10 +190,6 @@ export interface ValidatorListResponse {
     pagination: Pagination;
 }
 
-export interface ValidatorResponse {
-    result:     ValidatorInfo;
-}
-
 export interface ValidatorInfo {
     operatorAddress:           string;
     consensusNodeAddress:      string;
